pyproject.toml
[project]
name = "research-project"
version = "0.1.0"
description = "Research project template"
requires-python = ">=3.10"
dependencies = [
    "jax[cuda12]",
    "jax-metal>=0.1.1; sys_platform == 'darwin'",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "pytest>=7.3.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "scikit-image>=0.21.0",
    "requests>=2.31.0",
    "h5py>=3.9.0",
    "einops>=0.6.0",
    "jsonlines>=3.1.0",
    "umap-learn>=0.5.6",
    "numba>=0.58",
    "hdbscan>=0.8.41",
]

[dependency-groups]
dev = [
    "ipykernel>=6.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
